+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Logistic regression now works from 0 to 1 instead of -1 to 1.
 * Utility functions now import from `vowpal_platypus.utils`
 * Daemon functions now import from `vowpal_platypus.daemon`
+* `run_parallel` supports the pre v2 `run` function for running a manual function for each model on each core.
 
 **Minor/Technical Changes**